change amount to float

+++ src/main/java/pl/plantoplate/REST/dto/Request/AddShopProductRequest.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class AddShopProductRequest {
 
     private long id;
-    private int amount;
+    private float amount;
 }
